+++ orion/orion/plugins/__init__.py
@@ -1,5 +1,0 @@
-"""Orion plugin framework."""
-from orion.plugins.base import OrionPlugin, PluginContext
-from orion.plugins.manager import PluginManager, PluginRecord
-
-__all__ = ["OrionPlugin", "PluginContext", "PluginManager", "PluginRecord"]